Adding support for dependency injection via Microsoft.Extensions.DependencyInjection
This commit adds support for DI to the client application. By default, ReactiveUI utilizes Splat for the dependency injection framework. This works really well, but in order to support more robust scenarios we are taking advantage of Microsoft.Extensions.DependencyInjection instead. This commit also removes the sample `ViewLocator` that was included with the template project.