Request data support
This commit adds support for providing some request data for a request, although it is currently limited to sending `application/x-www-form-urlencoded`. For convenience, the request data textbox only appears for the request types "POST", "PATCH", and "PUT". In the future, I would like to allow users to override this if they want to send request data for other HTTP methods. I have also added a couple minor adjustments to the UI of the request builder to allow for a little bit more space between the stack panels.