Content type selector
This commit adds the ability to select a content type prior to sending request data. If a content type is not selected, the default continues to be `application/x-www-form-urlencoded`. The encoding is still set to UTF-8, which will probably be fine for most servers but may cause certain APIs to respond incorrectly (this can be addressed in a follow up).