Response header support
This commit adds support for displaying the response headers of a request, as well as some other minor enhancements such as the colorizing of response status codes. I have also cleaned up the layout a little bit, and changed the default window size of the request builder. Lastly, I have set it so the GET HTTP method is the default selected using the dropdown and I now prevent the form submission when the URL entered is invalid.