Workbook viewer
This commit implements a basic workbook viewer, which currently just draws out a fully-expanded tree view of all the folders and requests. It also added two interfaces (`IModel` and `ISerializableModel`) which assist in the implementation of nullable reference types in domain models. Since JSON deserialization does not guarantee that properties marked as non-nullable will indeed be this way at runtime, there needs to be a mechanism for allowing null values to be sanitized before being part of a domain model. `IModel{TModel, TSerializableModel}` and `ISerializableModel{TSerializableModel, TModel}` provide a single method each, which ensures that you have a method defined for converting to a domain model and to a serializable model from the respective implementations. While this is a trade off (since it requires a serializable class for each domain model that might be serialized), it affords us the ability to be confident that our domain models will not need defensive null checks throughout them.