Renaming and save workbook functionality
This commit adds the ability to rename and save a workbook. Unfortunately, this commit also removes the `IModel` and `ISerializableModel` stuff introduced in the previous commit. I realized that if we need to pass parameters into the `ToModel` and `ToSerializable` methods, we would be forced to break the interface or resort to having the methods take some hacky dictionary or loosely typed initialization structure.