New workbook folder dialog
This commit changes the new folder functionality to be a dialog instead of inline within the workbook window. This feels a little bit more proper and it also has the effect of solving the problem where it is not super obvious how to create a folder that is nested in a parent folder.