Introduce PathModel for request/folder paths
This commit implements a `PathModel` which takes the place of string-driven paths for requests and folders within workbooks. The primary goal of this is to allow for more flexibility around request and folder names. For example, this will allow folders to have slashes in their names. One of the other benefits of this change is that these paths are now much easier to manipulate using methods like `AddSegment` and `GetParent`.