Workbooks: Persist expand/collapse state into local app settings
This commit attempts to persist and load the expand/collapse state for folders in a workbook in the user's local app settings directory of their filesystem. The effect of this is that when the user loads a workbook from their filesystem, they will start back where they left off rather than having to re-expand/collapse the folders that they want to work on.