Adding a folder description
This commit adds the concept of a folder description. There is actually a good reason for this that is not so obvious: without having at least some metadata attached to folders, there is a strange quirk when attempting to commit workbooks to version control. Imagine you have an empty folder inside your workbook named "moo", you end up with the following filesystem paths:

- SomeWorkbook/
- SomeWorkbook/workbook.json
- SomeWorkbook/folders/
- SomeWorkbook/folders/moo/

The issue here is that Git does not actually track empty directories (or directories that only have empty directories inside of them). Adding a `folder.json` inside each folder solves this problem by introducing:
- SomeWorkbook/folders/moo/folder.json

In the future, we can take advantage of this file to store additional meta information for folders.